Section 4 - ContinuousTraining-Airflow-Composer/assignment/requirements.txt
scikit-learn==1.2.2
Flask==2.1.0
pandas==2.0.3
gunicorn==20.1.0
Werkzeug==2.0.3
numpy==1.25.2
pyarrow
pytest==8.3.2
category_encoders
requests==2.32.3
google-cloud-storage
google-cloud-bigquery==2.34.4
google-cloud-logging
dill